pyproject.toml
[tool.poetry]
name = "tasktracker"
version = "0.1.0"
description = ""
authors = ["romik <romik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.114.1"
pydantic = {extras = ["email"], version = "^2.9.1"}
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
pydantic-settings = "^2.5.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
redis = "^5.0.8"
celery = {extras = ["redis"], version = "^5.4.0"}
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
